Save builtins autodeploy key between restarts (#1161)
In order to authorize certain function calls in bultin services, the "builtins autodeploy" key is used. 

That key is a sister-key of the "management" key that's used to add aliases and such, with the important difference of keeping the private part in the node's reach, so node can use that key when auto-deploying builtin services.

This PR enables persistence of that key so that each builtin service need to adhere to a single administrative key.